completed auth flow

+++ app/Http/Controllers/Api/V1/Post/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -17,7 +18,10 @@ class PostController extends Controller
     {
     //    return PostResource::collection(Post::all());
     //    return PostResource::collection(Post::with('author')->get());
-       return PostResource::collection(Post::with('author')->paginate(2));
+
+      $user = request()->user();
+      $posts = $user->posts()->paginate();
+       return PostResource::collection($posts);
     }
 
     /**
@@ -36,7 +40,7 @@ class PostController extends Controller
 
     $data = $request->validated();
 
-    $data["author_id"]=1;
+    $data["author_id"]=$request->user()->id;
 
    $post = Post::create($data);
 
@@ -58,24 +62,22 @@ return response()->json(new PostResource($post),201);
     /**
      * Display the specified resource.
      */
-public function show(string $id)
+public function show(Post $post)
 {
-    $post = Post::find($id);
+    $user = request()->user();
 
-    if (!$post) {
-        return response()->json([
-            'message' => 'Post not found'
-        ], 404);
-    }
+    abort_if(!$user || $user->id !== $post->author_id, 403, 'Access Forbidden');
 
     return new PostResource($post);
 }
-
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, Post $post)
-    {
+    {   
+
+    abort_if(Auth::id() !== $post->author_id, 403, 'Access Forbidden');
+
         //
         $data = $request->validate([
         'title'=> 'required|string|max:100',
@@ -94,6 +96,8 @@ public function show(string $id)
      */
     public function destroy(Post $post)
     {
+
+    abort_if(Auth::id() !== $post->author_id, 403, 'Access Forbidden');
 
     $post->delete();
